Conflicts:
	sli/acceptance-tests/test/data/oauth_authentication_tokens.json
	sli/acceptance-tests/test/features/apiV1/associations/crud/assoc_crud.feature
	sli/acceptance-tests/test/features/apiV1/end_user_stories/CustomEntities/CustomEntities.feature
	sli/acceptance-tests/test/features/apiV1/end_user_stories/assessments/assessment.feature
	sli/acceptance-tests/test/features/ingestion/features/ingestion_securityEvent.feature
	sli/acceptance-tests/test/features/utils/sli_utils.rb
	sli/api/pom.xml
	sli/api/src/main/java/org/slc/sli/api/representation/AccessDeniedExceptionHandler.java
	sli/api/src/main/java/org/slc/sli/api/resources/security/SamlFederationResource.java
	sli/api/src/main/java/org/slc/sli/api/security/context/ContextValidator.java
	sli/api/src/main/java/org/slc/sli/api/security/context/WriteValidator.java
	sli/api/src/main/java/org/slc/sli/api/security/context/resolver/EdOrgHelper.java
	sli/api/src/test/java/org/slc/sli/api/resources/security/SamlFederationResourceTest.java
	sli/ingestion/ingestion-core/src/main/java/org/slc/sli/ingestion/processors/JobReportingProcessor.java

+++ sli/config/scripts/ingestion/slirp_reset.sh
@@ -14,6 +14,7 @@ NUMBER_OF_SERVERS=$1
 NUMBER_OF_TENANTS=$2
 
 TOTAL_INGESTION_SERVERS=3
+ISDB="slirpmongo99.slidev.org"
 
 if [ -z "$ING_LOG_DIR" ] ; then
   ING_LOG_DIR=/opt/logs
@@ -52,7 +53,7 @@ rm -r -f /ingestion/lz/inbound/*
 echo " ***** Dropping databases"
 ALL_DBS=`mongo --quiet --eval 'db.getMongo().getDBNames()' | sed -e 's/,/ /g'`
 for DB in $ALL_DBS ; do
-  if [ "$DB" != "test" -a "$DB" != "config" -a "$DB" != "local" ] ; then
+  if [ "$DB" != "test" -a "$DB" != "config" -a "$DB" != "local" -a "$DB" != "admin" ] ; then
     echo "Dropping database $DB"
     mongo $DB --quiet --eval 'db.dropDatabase()'
   fi
@@ -61,7 +62,7 @@ done
 echo " ***** Dropping ingestion_batch_job"
 ALL_DBS=`mongo $ISDB --quiet --eval 'db.getMongo().getDBNames()' | sed -e 's/,/ /g'`
 for DB in $ALL_DBS ; do
-  if [ "$DB" != "test" -a "$DB" != "config" -a "$DB" != "local" ] ; then
+  if [ "$DB" != "test" -a "$DB" != "config" -a "$DB" != "local" -a "$DB" != "admin" ] ; then
     echo "Dropping database $DB on $ISDB"
     mongo $ISDB/$DB --quiet --eval 'db.dropDatabase()'
   fi
@@ -111,6 +112,7 @@ TENANTS[6]="Tenant_4-State"
 TENANTS[7]="Tenant_5-State"
 for (( NUM=1; NUM<=$NUMBER_OF_TENANTS; NUM++ )) ; do
   TENANT=${TENANTS[$NUM]}
+  mkdir $LZ/inbound/$TENANT
   echo "***** Onboarding tenant #$NUM - $TENANT"
   echo "@purge" > /tmp/MainControlFile.ctl
   pushd /tmp
